module-one/pyproject.toml
[project]
name = "module-one"
version = "0.1.0"
description = "LangGraph Course Module One"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.50",
    "langchain-groq>=0.3.2",
    "langgraph>=0.3.25",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0"
]
[tool.ruff]
indent-width = 2

[tool.ruff.format]
indent-style = "tab"
